+++ src/index.ts
@@ -4,4 +4,4 @@ export { Statement } from './runtime/statement'
 export * as Index from './indexes'
 export * as Keyword from './keywords'
 export * as Element from './runtime/element'
-export { convert, parseExpr } from './lang/js/converter'
+export { convert, parseExpr } from './tool/converter'
