--- src/lang/js/expression.ts
@@ -1,5 +1,0 @@
-import * as calcium from '../../api/index'
-
-export type Reference = [string, ...any[]]
-
-export type Expression = calcium.Element | Reference
